Zad2c (3/4) Poprawa returnu w StatisticsHelper.cs aby zwracał double

+++ apbd-cw1-git-s32948_v2/StatisticsHelper.cs
@@ -9,6 +9,6 @@ public class StatisticsHelper
         {
             sum += n;
         }
-        return sum / numbers.Length;
+        return  (double) sum / numbers.Length;
     }
 }
